+++ src/app/core/local-storage.service.ts
@@ -10,7 +10,7 @@ export class LocalStorageService {
     localStorage.setItem(key, value);
   }
 
-  getItem(key: string): string {
+  getItem(key: string): string | null {
     return localStorage.getItem(key);
   }